Add new query APIs
  * Move endpoints from stats to query resource
  * Add package type parameter to endpoints API
  * Add storekeys API with package type parameter
