+++ src/main/java/org/commonjava/indy/service/repository/controller/StatsController.java
@@ -16,24 +16,10 @@
 package org.commonjava.indy.service.repository.controller;
 
 import org.commonjava.indy.service.repository.data.StoreDataManager;
-import org.commonjava.indy.service.repository.exception.IndyDataException;
-import org.commonjava.indy.service.repository.exception.IndyWorkflowException;
-import org.commonjava.indy.service.repository.model.ArtifactStore;
-import org.commonjava.indy.service.repository.model.StoreKey;
-import org.commonjava.indy.service.repository.model.dto.EndpointView;
-import org.commonjava.indy.service.repository.model.dto.EndpointViewListing;
 import org.commonjava.indy.service.repository.model.version.Versioning;
-import org.commonjava.indy.service.repository.util.JaxRsUriFormatter;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-import static javax.ws.rs.core.Response.Status.INTERNAL_SERVER_ERROR;
 
 @ApplicationScoped
 public class StatsController
@@ -54,58 +40,6 @@ public class StatsController
     public Versioning getVersionInfo()
     {
         return versioning;
-    }
-
-    public EndpointViewListing getEndpointsListing( final String baseUri, final JaxRsUriFormatter uriFormatter )
-            throws IndyWorkflowException
-    {
-        final List<ArtifactStore> stores;
-        try
-        {
-            stores = new ArrayList<>( dataManager.getAllArtifactStores() );
-        }
-        catch ( final IndyDataException e )
-        {
-            throw new IndyWorkflowException( INTERNAL_SERVER_ERROR.getStatusCode(),
-                                             "Failed to retrieve all endpoints: {}", e, e.getMessage() );
-        }
-
-        final List<EndpointView> points = new ArrayList<>();
-        for ( final ArtifactStore store : stores )
-        {
-            final StoreKey key = store.getKey();
-            final String resourceUri = uriFormatter.formatAbsolutePathTo( baseUri, "content", key.getPackageType(),
-                                                                          key.getType().singularEndpointName(),
-                                                                          key.getName() );
-
-            final EndpointView point = new EndpointView( store, resourceUri );
-            if ( !points.contains( point ) )
-            {
-                points.add( point );
-            }
-        }
-
-        return new EndpointViewListing( points );
-    }
-
-    public Map<String, List<String>> getAllStoreKeys()
-            throws IndyWorkflowException
-    {
-        final List<ArtifactStore> stores;
-        try
-        {
-            final Map<String, List<String>> result = new HashMap<>();
-            stores = new ArrayList<>( dataManager.getAllArtifactStores() );
-            List<String> items = stores.stream().map( s-> s.getKey().toString() ).collect( Collectors.toList() );
-            result.put("items", items);
-            return result;
-        }
-        catch ( final IndyDataException e )
-        {
-            throw new IndyWorkflowException( INTERNAL_SERVER_ERROR.getStatusCode(),
-                                             "Failed to retrieve all store keys: {}", e, e.getMessage() );
-        }
-
     }
 
 }
